rename: table enum 파일 이름을 변경한다
