+++ src/database/types/enums/table.enum.ts
